WIP: universal root mod support
~ moved the plugin two folders back, from "plugins\basic_games\games\blah.py" to "plugins\blah.py"
  this apparently allows us to provide not just a game plugin but also an "installer" plugin.
+ added super minimal DarktideModInstaller class
  the initial plan will be to try and preemptively "fix" the file tree of any mod the user wants to install, and the user will still be able to manipulate the file tree as usual.

In order for this to ever possibly work, the default "GameDataPath" has been set to a nonsensical folder name (inspired by the Subnautica plugin) in the hope that it will have no effect on the game whatsoever.
The goal is for us to be responsible for ALL the mapping ourselves.
Then it would only be a matter of deciding if a mod is a root mod or not, and map it accordingly.
~ adjusted `build_custom_mappings()` to simply map all mods either relative to "<Game>/" or "<Game>/mods/", with DML still being hardcoded as the only root mod for now.

Various other paths had to be adjusted to account for the nonsensical GameDataPath change.
e.g. `self.dataDirectory()` became `Path(self.gameDirectory().absolutePath()) / "mods"`

+ dump all mappings in debug info

~ adjusted `identify_mod()` to attempt to detect root mods.
  just checks for presence of "mods/" dir or ".is_root_mod" file/dir...
  saves the result in the mod's settings, which is persisted in the mod's "meta.ini" file.
  this would probably be the thing to use to decide where to map a mod.

~ the custom hardcoded post-installation step for DML has been disabled for now